CAUSEWAY-3825: initial draft of TabularModel
