+++ commons/src/main/java/module-info.java
@@ -26,6 +26,7 @@ module org.apache.causeway.commons {
     exports org.apache.causeway.commons.handler;
     exports org.apache.causeway.commons.having;
     exports org.apache.causeway.commons.io;
+    exports org.apache.causeway.commons.tabular;
     // internals exported as well
     exports org.apache.causeway.commons.internal;
     exports org.apache.causeway.commons.internal.assertions;
